refactor: improve structure and naming
